Move util tests to dedicated file
