+++ integrations/amazon_bedrock/tests/test_chat_generator.py
@@ -9,11 +9,6 @@ from haystack.dataclasses import ChatMessage, ChatRole, StreamingChunk
 from haystack.tools import Tool
 
 from haystack_integrations.components.generators.amazon_bedrock import AmazonBedrockChatGenerator
-from haystack_integrations.components.generators.amazon_bedrock.chat.utils import (
-    _format_tools,
-    _parse_completion_response,
-    _parse_streaming_response,
-)
 
 CLASS_TYPE = "haystack_integrations.components.generators.amazon_bedrock.chat.chat_generator.AmazonBedrockChatGenerator"
 MODELS_TO_TEST = [
@@ -270,136 +265,6 @@ class TestAmazonBedrockChatGenerator:
             ] = tool.outputs_to_state
 
         assert pipeline_dict == expected_dict
-
-    def test_format_tools(self, tools):
-        formatted_tool = _format_tools(tools[:1])
-        assert formatted_tool == {
-            "tools": [
-                {
-                    "toolSpec": {
-                        "name": "weather",
-                        "description": "useful to determine the weather in a given location",
-                        "inputSchema": {
-                            "json": {
-                                "type": "object",
-                                "properties": {"city": {"type": "string"}},
-                                "required": ["city"],
-                            }
-                        },
-                    }
-                }
-            ],
-        }
-
-    def test_extract_replies_from_response(self, mock_boto3_session):
-        """
-        Test that extract_replies_from_response correctly processes both text and tool use responses
-        """
-        model = "anthropic.claude-3-5-sonnet-20240620-v1:0"
-        # Test case 1: Simple text response
-        text_response = {
-            "output": {"message": {"role": "assistant", "content": [{"text": "This is a test response"}]}},
-            "stopReason": "complete",
-            "usage": {"inputTokens": 10, "outputTokens": 20, "totalTokens": 30},
-        }
-
-        replies = _parse_completion_response(text_response, model)
-        assert len(replies) == 1
-        assert replies[0].text == "This is a test response"
-        assert replies[0].role == ChatRole.ASSISTANT
-        assert replies[0].meta["model"] == model
-        assert replies[0].meta["finish_reason"] == "complete"
-        assert replies[0].meta["usage"] == {"prompt_tokens": 10, "completion_tokens": 20, "total_tokens": 30}
-
-        # Test case 2: Tool use response
-        tool_response = {
-            "output": {
-                "message": {
-                    "role": "assistant",
-                    "content": [{"toolUse": {"toolUseId": "123", "name": "test_tool", "input": {"key": "value"}}}],
-                }
-            },
-            "stopReason": "tool_call",
-            "usage": {"inputTokens": 15, "outputTokens": 25, "totalTokens": 40},
-        }
-
-        replies = _parse_completion_response(tool_response, model)
-        assert len(replies) == 1
-        tool_content = replies[0].tool_call
-        assert tool_content.id == "123"
-        assert tool_content.tool_name == "test_tool"
-        assert tool_content.arguments == {"key": "value"}
-        assert replies[0].meta["finish_reason"] == "tool_call"
-        assert replies[0].meta["usage"] == {"prompt_tokens": 15, "completion_tokens": 25, "total_tokens": 40}
-
-        # Test case 3: Mixed content response
-        mixed_response = {
-            "output": {
-                "message": {
-                    "role": "assistant",
-                    "content": [
-                        {"text": "Let me help you with that. I'll use the search tool to find the answer."},
-                        {"toolUse": {"toolUseId": "456", "name": "search_tool", "input": {"query": "test"}}},
-                    ],
-                }
-            },
-            "stopReason": "complete",
-            "usage": {"inputTokens": 25, "outputTokens": 35, "totalTokens": 60},
-        }
-
-        replies = _parse_completion_response(mixed_response, model)
-        assert len(replies) == 1
-        assert replies[0].text == "Let me help you with that. I'll use the search tool to find the answer."
-        tool_content = replies[0].tool_call
-        assert tool_content.id == "456"
-        assert tool_content.tool_name == "search_tool"
-        assert tool_content.arguments == {"query": "test"}
-
-    def test_process_streaming_response(self, mock_boto3_session):
-        """
-        Test that process_streaming_response correctly handles streaming events and accumulates responses
-        """
-        model = "anthropic.claude-3-5-sonnet-20240620-v1:0"
-
-        streaming_chunks = []
-
-        def test_callback(chunk: StreamingChunk):
-            streaming_chunks.append(chunk)
-
-        # Simulate a stream of events for both text and tool use
-        events = [
-            {"contentBlockStart": {"start": {"text": ""}}},
-            {"contentBlockDelta": {"delta": {"text": "Let me "}}},
-            {"contentBlockDelta": {"delta": {"text": "help you."}}},
-            {"contentBlockStop": {}},
-            {"contentBlockStart": {"start": {"toolUse": {"toolUseId": "123", "name": "search_tool"}}}},
-            {"contentBlockDelta": {"delta": {"toolUse": {"input": '{"query":'}}}},
-            {"contentBlockDelta": {"delta": {"toolUse": {"input": '"test"}'}}}},
-            {"contentBlockStop": {}},
-            {"messageStop": {"stopReason": "complete"}},
-            {"metadata": {"usage": {"inputTokens": 10, "outputTokens": 20, "totalTokens": 30}}},
-        ]
-
-        replies = _parse_streaming_response(events, test_callback, model)
-
-        # Verify streaming chunks were received for text content
-        assert len(streaming_chunks) == 2
-        assert streaming_chunks[0].content == "Let me "
-        assert streaming_chunks[1].content == "help you."
-
-        # Verify final replies
-        assert len(replies) == 2
-        # Check text reply
-        assert replies[0].text == "Let me help you."
-        assert replies[0].meta["model"] == model
-        assert replies[0].meta["finish_reason"] == "complete"
-        assert replies[0].meta["usage"] == {"prompt_tokens": 10, "completion_tokens": 20, "total_tokens": 30}
-
-        # Check tool use reply
-        tool_content = replies[1].tool_call
-        assert tool_content.id == "123"
-        assert tool_content.tool_name == "search_tool"
-        assert tool_content.arguments == {"query": "test"}
 
 
 class TestAmazonBedrockChatGeneratorInference:
